Refactor render to a component
